exception controller test added
exception controller test added

+++ ApiApp/src/main/java/gradlePostmanTest/ApiApp/ProductServiceController.java
@@ -1,6 +1,8 @@
 package gradlePostmanTest.ApiApp;
 
+import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 import org.springframework.http.HttpStatus;
@@ -10,6 +12,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
+
+
 
 @RestController
 public class ProductServiceController {
@@ -24,17 +28,19 @@ public class ProductServiceController {
 		Product almond = new Product();
 		almond.setId("2");
 		almond.setName("Almond");
-		productRepo.put(honey.getId(), almond);
+		productRepo.put(almond.getId(), almond);
 	}
 
 	@RequestMapping(value = "/products/{id}", method = RequestMethod.DELETE)
 	public ResponseEntity<Object> delete(@PathVariable("id") String id) {
+		if (!productRepo.containsKey(id)) throw new ProductNotFoundException(); 
 		productRepo.remove(id);
-		return new ResponseEntity<Object>("Product is deleted sucessfully", HttpStatus.OK);
+		return new ResponseEntity<Object>("Product was deleted sucessfully", HttpStatus.OK);
 	}
 
 	@RequestMapping(value = "/products/{id}", method = RequestMethod.PUT)
 	public ResponseEntity<Object> updateProduct(@PathVariable("id") String id, @RequestBody Product product) {
+		if (!productRepo.containsKey(id)) throw new ProductNotFoundException();
 		productRepo.remove(id);
 		product.setId(id);
 		productRepo.put(id, product);
@@ -49,6 +55,14 @@ public class ProductServiceController {
 
 	@RequestMapping(value = "/products")
 	public ResponseEntity<Object> getProduct() {
-		return new ResponseEntity<>(productRepo.values(), HttpStatus.OK);
+		List<Product> productos = new ArrayList<Product>(productRepo.values());
+		return new ResponseEntity<>(productos, HttpStatus.OK);
+	}
+	
+	@RequestMapping(value = "/products/{id}")
+	public  ResponseEntity<Object> getProductById(@PathVariable("id") String id){
+		if (!productRepo.containsKey(id)) throw new ProductNotFoundException();
+		
+		return new ResponseEntity<Object>(productRepo.get(id), HttpStatus.OK);
 	}
 }
